servidorUbuntu2/backend/requirements.txt
asgiref==3.11.1
Django==6.0.3
django-filter==25.2
djangorestframework==3.16.1
psycopg2-binary==2.9.11
sqlparse==0.5.5
python-dotenv==1.2.2
